Remove low-quality unit test
